common/client distinction (sorta, need to review)
