--- src/main/java/dev/rndmorris/somberassembly/ClientProxy.java
@@ -1,8 +1,0 @@
-package dev.rndmorris.somberassembly;
-
-public class ClientProxy extends CommonProxy {
-
-    // Override CommonProxy methods here, if you want a different behaviour on the client (e.g. registering renders).
-    // Don't forget to call the super methods as well.
-
-}
