+++ src/main/java/dev/rndmorris/somberassembly/client/ClientProxy.java
@@ -1,0 +1,11 @@
+package dev.rndmorris.somberassembly.client;
+
+import dev.rndmorris.somberassembly.common.CommonProxy;
+
+public class ClientProxy extends CommonProxy
+{
+
+    // Override CommonProxy methods here, if you want a different behaviour on the client (e.g. registering renders).
+    // Don't forget to call the super methods as well.
+
+}
