--- src/main/java/dev/rndmorris/somberassembly/data/Tuple2.java
@@ -1,6 +1,0 @@
-package dev.rndmorris.somberassembly.data;
-
-import com.github.bsideup.jabel.Desugar;
-
-@Desugar
-public record Tuple2<T1, T2> (T1 value1, T2 value2) {}
