+++ src/main/java/dev/rndmorris/somberassembly/utils/Tuple2.java
@@ -1,0 +1,6 @@
+package dev.rndmorris.somberassembly.utils;
+
+import com.github.bsideup.jabel.Desugar;
+
+@Desugar
+public record Tuple2<T1, T2> (T1 value1, T2 value2) {}
